fix: rename getConfig as `getMfeConfig` and move it in own files
