+++ .vscode/settings.json
@@ -1,0 +1,16 @@
+{
+  // Enable ESLint
+  "eslint.validate": [
+    "javascript",
+    "javascriptreact",
+    "typescript",
+    "typescriptreact"
+  ],
+  // be sure vscode always uses TS version in local project
+  "typescript.tsdk": "node_modules/typescript/lib",
+  // format on save
+  // "editor.formatOnSave": true,
+  "editor.codeActionsOnSave": {
+    "source.fixAll": "explicit"
+  }
+}
